Upgraded frontend dependencies.
