+++ src/Frontend/package.json
@@ -16,7 +16,7 @@
     "classnames": "^2.5.1",
     "fast-deep-equal": "^3.1.3",
     "material-icons": "^1.13.12",
-    "pokeaclient": "^0.5.0",
+    "pokeaclient": "^0.5.1",
     "preact": "^10.26.6",
     "wouter": "^3.3.5"
   },
