+++ src/Frontend/src/pages/propertyEditor/components/PropertyTextbox.tsx
@@ -6,7 +6,6 @@ export function getPropertyFieldValue(value: any, type: GamePropertyType | null)
 	if (type === "bitArray") {
 		return value.map((x: boolean) => x ? "1" : "0").join("") ?? "";
 	}
-	// @ts-expect-error
 	if (type === "byteArray") {
 		console.log(value);
 		return value.map((x:number) => x.toString(16).toUpperCase().padStart(2, "0")).join(" ");
